Ladybird/WebView: Move WebSocket stuff to its own files
